+++ include/Vector2f.h
@@ -4,40 +4,40 @@
 
 class Vector2f {
     private:
-	double x;
-	double y;
+	float x;
+	float y;
 
     public:
 	Vector2f();
-	Vector2f(double a, double b);
+	Vector2f(float a, float b);
 
-	void setX(double a) noexcept;
-	void setY(double b) noexcept;
-	double getX() const noexcept;
-	double getY() const noexcept;
-	std::array<double, 2> get() const noexcept;
+	void setX(float a) noexcept;
+	void setY(float b) noexcept;
+	float getX() const noexcept;
+	float getY() const noexcept;
+	std::array<float, 2> get() const noexcept;
 
-	double length() const noexcept;
-	double dot(const Vector2f &v) noexcept;
+	float length() const noexcept;
+	float dot(const Vector2f &v) noexcept;
 	Vector2f normalize();
 
 	Vector2f operator+(const Vector2f &v);
 	Vector2f operator-(const Vector2f &v);
 	Vector2f operator*(const Vector2f &v);
 	Vector2f operator/(const Vector2f &v);
-	Vector2f operator+(const double r);
-	Vector2f operator-(const double r);
-	Vector2f operator*(const double r);
-	Vector2f operator/(const double r);
+	Vector2f operator+(const float r);
+	Vector2f operator-(const float r);
+	Vector2f operator*(const float r);
+	Vector2f operator/(const float r);
 
 	Vector2f &operator+=(const Vector2f &v);
 	Vector2f &operator-=(const Vector2f &v);
 	Vector2f &operator*=(const Vector2f &v);
 	Vector2f &operator/=(const Vector2f &v);
-	Vector2f &operator+=(const double r);
-	Vector2f &operator-=(const double r);
-	Vector2f &operator*=(const double r);
-	Vector2f &operator/=(const double r);
+	Vector2f &operator+=(const float r);
+	Vector2f &operator-=(const float r);
+	Vector2f &operator*=(const float r);
+	Vector2f &operator/=(const float r);
 
-	Vector2f rotate(double degrees) const noexcept;
+	Vector2f rotate(float degrees) const noexcept;
 };
